chore: rename env.py to environment_variables.py
